Move basic primitive impls to new module
Create a new private `consensus::impls` module and move the
`Encodable`/`Decodable` implemenations for basic primitives into it.
